Make modutils file, and rename get_attributes_names func
